decrating: absorb shipper-store into shipper::state::store
Move shipper-store crate (2816 LOC) into crates/shipper/src/state/store/
as a set of crate-private modules, split into mod.rs + fs.rs +
tests.rs + snapshot_tests.rs + path_edge_case_tests.rs. Delete the stale
386-LOC duplicate at crates/shipper/src/store.rs and the 1-LOC re-export
shim at store_micro.rs. Remove shipper-store from the workspace members
and from shipper's dependencies. Drop the micro-store feature flag from
both shipper and shipper-cli.

The StateStore trait stays as a trait -- multiple impls (fs, future
cloud, mocks for tests). The FileStore filesystem impl keeps its atomic
write semantics via the existing crate::state helpers.

The physical files live under state/store/ to match the layered
architecture (layer 3: state). The public path shipper::store is
preserved via a #[path] attribute in lib.rs for backward compatibility
with integration tests and downstream consumers.

NOTE: This PR was intended to be stacked on the events+state absorption
PR (feature/decrating-absorb-events-state) so that state/ could be a
first-class folder with state/mod.rs and state/CLAUDE.md. That PR was
not pushed within the coordination window, so this PR branched directly
from feature/decrating-phase1-scaffold and uses #[path] instead of a
state/mod.rs entry. When events+state lands, a follow-up can convert
state.rs -> state/mod.rs and add `pub(crate) mod store;` there.

All 28 snapshot files moved from crates/shipper-store/src/snapshots to
crates/shipper/src/state/store/snapshots and renamed to match the new
module path (shipper__store__snapshot_tests__*).

Per docs/decrating-plan.md Phase 2.
