+++ crates/shipper-cli/Cargo.toml
@@ -44,7 +44,6 @@ micro-webhook = ["shipper/micro-webhook"]
 micro-types = ["shipper/micro-types"]
 micro-config = ["shipper/micro-config"]
 micro-state = ["shipper/micro-state"]
-micro-store = ["shipper/micro-store"]
 micro-all = ["shipper/micro-all"]
 
 [dev-dependencies]
