--- crates/shipper-store/CLAUDE.md
@@ -1,29 +1,0 @@
-# CLAUDE.md
-
-This file provides agent-specific guidance for working in crate shipper-store.
-
-## Scope
-
-- Crate: shipper-store
-- Path: crates/shipper-store
-- Workspace root: h:\Code\Rust\shipper
-- Primary entry: src/lib.rs
-
-## Useful commands
-
-```bash
-cargo check -p shipper-store
-cargo test -p shipper-store
-cargo test -p shipper-store --all-features
-cargo fmt -p shipper-store
-cargo clippy -p shipper-store --all-targets --all-features -- -D warnings
-```
-
-## Context
-
-- Keep changes small and targeted to the crate’s existing abstractions.
-- Preserve public API compatibility unless the request explicitly asks for breaking changes.
-- When touching serialization or state formats, update tests and related snapshots in the same crate.
-- Prefer using existing fixtures and helpers rather than introducing inline test data.
-
-For full workspace guidance, see [../../CLAUDE.md](H:\Code\Rust\shipper\CLAUDE.md).
