--- crates/shipper-store/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "shipper-store"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-description = "State store abstraction for shipper"
-repository = "https://github.com/EffortlessMetrics/shipper"
-documentation = "https://docs.rs/shipper-store"
-keywords = ["store", "state", "persistence"]
-categories = ["development-tools"]
-
-[dependencies]
-anyhow = "1.0"
-serde.workspace = true
-serde_json = "1.0"
-chrono = { version = "0.4.44", features = ["serde"] }
-shipper-events = { path = "../shipper-events" }
-shipper-types = { path = "../shipper-types" }
-shipper-state = { path = "../shipper-state" }
-shipper-schema = { path = "../shipper-schema" }
-
-[dev-dependencies]
-tempfile = "3"
-proptest.workspace = true
-insta = { workspace = true }
-
-[lints]
-workspace = true
