+++ crates/shipper/Cargo.toml
@@ -31,7 +31,6 @@ shipper-types = { path = "../shipper-types" }
 shipper-config = { path = "../shipper-config" }
 shipper-state = { path = "../shipper-state", optional = true }
 shipper-config-runtime = { path = "../shipper-config-runtime" }
-shipper-store = { path = "../shipper-store", optional = true }
 shipper-duration = { path = "../shipper-duration" }
 shipper-schema = { path = "../shipper-schema" }
 shipper-cargo-failure = { path = "../shipper-cargo-failure" }
@@ -76,9 +75,8 @@ micro-webhook = []
 micro-types = ["micro-webhook", "micro-encrypt", "micro-config"]
 micro-config = ["micro-types", "micro-storage"]
 micro-state = ["dep:shipper-state", "micro-types", "micro-environment", "micro-encrypt"]
-micro-store = ["dep:shipper-store", "micro-types", "micro-events", "micro-state"]
 micro-parallel = ["micro-types", "micro-plan", "micro-registry", "dep:shipper-engine-parallel"]
-micro-all = ["micro-auth", "micro-git", "micro-events", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-registry", "micro-webhook", "micro-types", "micro-config", "micro-state", "micro-store", "micro-parallel"]
+micro-all = ["micro-auth", "micro-git", "micro-events", "micro-encrypt", "micro-environment", "micro-storage", "micro-cargo", "micro-plan", "micro-registry", "micro-webhook", "micro-types", "micro-config", "micro-state", "micro-parallel"]
 
 [lints]
 workspace = true
