--- crates/shipper/src/store_micro.rs
@@ -1,1 +1,0 @@
-pub use shipper_store::*;
